requirements/base.txt
Django==5.2.3
psycopg2-binary==2.9.10
python-environ==0.4.54
django-cors-headers==4.7.0
django-jazzmin==3.0.1
drf-yasg==1.21.10
djangorestframework==3.16.0
django-filter==25.1
django-modeltranslation==0.19.15
django-redis==6.0.0
django-recaptcha==3.0.0
nplusone==1.0.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.3
pillow==11.2.1
djangorestframework-simplejwt==5.2.2
httpx==0.27.0
adrf==0.1.9
payme-pkg==3.0.27
click-pkg==0.15
django-tinymce==3.7.1
django-celery-beat==2.8.1
requests==2.31.0
django-ckeditor